Add option to not manage the Sphinx service
Previously, the Sphinx service was being managed regardless of how the main
sphinx class was declared. This commit adds the ability to pass a $service_enable
parameter, as well as exposes the $sphinx_activate_service parameter to the main
sphinx class (that way you can pass in the parameter without having to directly
modify the variable in the params class, thus making the class more modular).
By exposing these parameters, it enables the ability to disable managing the
Sphinx service (should it be controlled by other means).